Large overhaul for addition of reverse proxy
Added reverse proxy to route traffit from subdomains to the
appropriate services.
Proxy checks for authentication token when routing traffic to the
FHIR server.
Removed unnecessary OpenID config options for OpenID and CareShare
URLs.
Split error messages for "unauthenticated" and "unauthorized" users
(now a user that tries to do an admin action will get a 403 error
instead of a 401 error).